requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Pillow>=10.1.0
pillow-heif==0.21.0
qrcode==7.3.1
gunicorn==20.1.0
jieba==0.42.1
requests==2.27.1
selenium==3.141.0
python-dotenv==0.19.0
